backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
passlib[argon2]
python-jose
pydantic
alembic
pytest
pyjwt
pwdlib[argon2]
pydantic_settings
pydantic[email]